1.18.2 update

+++ src/main/java/ga/gamer234emp/obv/procedures/BLKTNBTProcedure.java
@@ -9,9 +9,9 @@ import net.minecraft.core.BlockPos;
 
 public class BLKTNBTProcedure {
 	public static void execute(LevelAccessor world, double x, double y, double z, BlockState blockstate) {
-		if ((blockstate.getBlock().getStateDefinition().getProperty("age")instanceof IntegerProperty _ip ? blockstate.getValue(_ip) : -1) == 1) {
+		if ((blockstate.getBlock().getStateDefinition().getProperty("age") instanceof IntegerProperty _ip ? blockstate.getValue(_ip) : -1) == 1) {
 			if (!world.isClientSide()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				BlockEntity _blockEntity = world.getBlockEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_blockEntity != null)
@@ -19,11 +19,11 @@ public class BLKTNBTProcedure {
 				if (world instanceof Level _level)
 					_level.sendBlockUpdated(_bp, _bs, _bs, 3);
 			}
-		} else if ((blockstate.getBlock().getStateDefinition().getProperty("age")instanceof IntegerProperty _ip
+		} else if ((blockstate.getBlock().getStateDefinition().getProperty("age") instanceof IntegerProperty _ip
 				? blockstate.getValue(_ip)
 				: -1) == 2) {
 			if (!world.isClientSide()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				BlockEntity _blockEntity = world.getBlockEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_blockEntity != null)
@@ -31,11 +31,11 @@ public class BLKTNBTProcedure {
 				if (world instanceof Level _level)
 					_level.sendBlockUpdated(_bp, _bs, _bs, 3);
 			}
-		} else if ((blockstate.getBlock().getStateDefinition().getProperty("age")instanceof IntegerProperty _ip
+		} else if ((blockstate.getBlock().getStateDefinition().getProperty("age") instanceof IntegerProperty _ip
 				? blockstate.getValue(_ip)
 				: -1) == 3) {
 			if (!world.isClientSide()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				BlockEntity _blockEntity = world.getBlockEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_blockEntity != null)
@@ -43,11 +43,11 @@ public class BLKTNBTProcedure {
 				if (world instanceof Level _level)
 					_level.sendBlockUpdated(_bp, _bs, _bs, 3);
 			}
-		} else if ((blockstate.getBlock().getStateDefinition().getProperty("age")instanceof IntegerProperty _ip
+		} else if ((blockstate.getBlock().getStateDefinition().getProperty("age") instanceof IntegerProperty _ip
 				? blockstate.getValue(_ip)
 				: -1) == 4) {
 			if (!world.isClientSide()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				BlockEntity _blockEntity = world.getBlockEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_blockEntity != null)
@@ -55,11 +55,11 @@ public class BLKTNBTProcedure {
 				if (world instanceof Level _level)
 					_level.sendBlockUpdated(_bp, _bs, _bs, 3);
 			}
-		} else if ((blockstate.getBlock().getStateDefinition().getProperty("age")instanceof IntegerProperty _ip
+		} else if ((blockstate.getBlock().getStateDefinition().getProperty("age") instanceof IntegerProperty _ip
 				? blockstate.getValue(_ip)
 				: -1) == 0) {
 			if (!world.isClientSide()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				BlockEntity _blockEntity = world.getBlockEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_blockEntity != null)
