+++ src/main/java/ga/gamer234emp/obv/procedures/LargeBatteryDestroyedProcedure.java
@@ -5,6 +5,6 @@ import net.minecraft.core.BlockPos;
 
 public class LargeBatteryDestroyedProcedure {
 	public static void execute(LevelAccessor world, double x, double y, double z) {
-		world.destroyBlock(new BlockPos((int) x, (int) (y + 1), (int) z), false);
+		world.destroyBlock(new BlockPos(x, y + 1, z), false);
 	}
 }
