+++ src/main/java/ga/gamer234emp/obv/procedures/OrangeUiLightProcedure.java
@@ -13,13 +13,13 @@ public class OrangeUiLightProcedure {
 					return blockEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "temperature") < 1199 && (new Object() {
+		}.getValue(world, new BlockPos(x, y, z), "temperature") < 1199 && (new Object() {
 			public String getValue(LevelAccessor world, BlockPos pos, String tag) {
 				BlockEntity blockEntity = world.getBlockEntity(pos);
 				if (blockEntity != null)
 					return blockEntity.getTileData().getString(tag);
 				return "";
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "isActive")).equals("yes");
+		}.getValue(world, new BlockPos(x, y, z), "isActive")).equals("yes");
 	}
 }
