+++ src/main/java/ga/gamer234emp/obv/procedures/ThreeRodReturnProcedure.java
@@ -13,6 +13,6 @@ public class ThreeRodReturnProcedure {
 					return blockEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "numOfUrandia") >= 3;
+		}.getValue(world, new BlockPos(x, y, z), "numOfUrandia") >= 3;
 	}
 }
